+++ R/ui_helpers.R
@@ -81,7 +81,7 @@ app_brand_title <- function(version) {
 }
 
 app_stylesheet_link <- function(version) {
-  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-independent-heights-4"))
+  tags$link(rel = "stylesheet", type = "text/css", href = paste0("style.css?v=", version, "-reliability-factors-15"))
 }
 
 app_script_link <- function(version) {